requirements.txt
aniso8601==8.0.0
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
Click==7.0
Flask==1.1.1
Flask-Injector==0.12.2
flask-restplus==0.13.0
importlib-metadata==1.5.0
injector==0.18.3
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.4
PyJWT==1.7.1
pylint==2.4.4
pyrsistent==0.15.7
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.13
sqlparse==0.3.0
typed-ast==1.4.1
typing-extensions==3.7.4.1
Werkzeug==0.16.1
wrapt==1.11.2
zipp==3.0.0
